agents/claude-adk/pyproject.toml
[project]
name = "claude-adk-agent"
version = "0.1.0"
description = "Python agent development environment scaffold"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.53.0",
  "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.6.0"
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
